moved custom view model class to proper package
